demo/ann_requirements_demo.txt
Flask
Flask-SQLAlchemy
scikit-learn
pandas
numpy
joblib
gunicorn

tensorflow-cpu
tf-keras
transformers
sentencepiece
torch